Small music improvement

+++ bot/cogs/music.py
@@ -202,8 +202,9 @@ class MusicController(ui.View):
 
     async def on_timeout(self):
         await self.player.controller_message.edit(view=None)
-        await self.player.controller_message.delete()
-        del self.player.controller_message
+        if hasattr(self.player, 'controller_message'):
+            await self.player.controller_message.delete()
+            del self.player.controller_message
 
     async def interaction_check(self, interaction: Interaction) -> bool:
         if interaction.user != self.ctx.author:
